+++ src/panorama/utils/utils.ts
@@ -26,18 +26,23 @@ export function GetOpenMemberUrl(): string {
   return $.Localize('#player_member_ship_url') + GetLocalPlayerSteamAccountID();
 }
 
-// 自定义键位绑定
-// export function AddKeyBind(
-//   keyName: string,
-//   keydownCallback?: () => void,
-//   keyupCallback?: () => void,
-// ) {
-//   const command = `${keyName}${Date.now()}`;
-//   Game.CreateCustomKeyBind(command, `+${command}`);
-//   if (keydownCallback) {
-//     Game.AddCommand(`+${command}`, keydownCallback, '', 1 << 32);
-//   }
-//   if (keyupCallback) {
-//     Game.AddCommand(`-${command}`, keyupCallback, '', 1 << 32);
-//   }
-// }
+/**
+ * 添加自定义键位绑定
+ * @param keyName 键位名称
+ * @param keydownCallback 按下键位时的回调函数
+ * @param keyupCallback 松开键位时的回调函数
+ */
+export function AddKeyBind(
+  keyName: string,
+  keydownCallback?: () => void,
+  keyupCallback?: () => void,
+) {
+  const command = `${keyName}${Date.now()}`;
+  Game.CreateCustomKeyBind(command, `+${command}`);
+  if (keydownCallback) {
+    Game.AddCommand(`+${command}`, keydownCallback, '', 1 << 32);
+  }
+  if (keyupCallback) {
+    Game.AddCommand(`-${command}`, keyupCallback, '', 1 << 32);
+  }
+}
